Xcode 5.1 compatible

+++ MKStoreManager.m
@@ -259,7 +259,7 @@ static MKStoreManager* _sharedStoreManager;
 - (BOOL) removeAllKeychainData {
   
   NSMutableArray *productsArray = [MKStoreManager allProducts];
-  int itemCount = productsArray.count;
+  NSUInteger itemCount = productsArray.count;
   NSError *error;
   
   //loop through all the saved keychain data and remove it
@@ -430,7 +430,7 @@ static MKStoreManager* _sharedStoreManager;
   if ([SKPaymentQueue canMakePayments])
 	{
         NSArray *allIds = [self.purchasableObjects valueForKey:@"productIdentifier"];
-        int index = [allIds indexOfObject:productId];
+        NSUInteger index = [allIds indexOfObject:productId];
     
         if(index == NSNotFound) {
             if (self.onTransactionCancelled) {
